+++ CHANGELOG.md
@@ -16,6 +16,8 @@
         - Change text on /reports to match lower down (fix translation).
         - Ensure all reports graph can't dip downward. #1956
         - Fix error sending `requires_inspection` reports. #1961
+    - Admin improvements:
+        - Admin can anonymize/hide all a user's reports. #1942 #1943
     - UK:
         - Lazy load images in the footer.